gpui button, moved elements to their own module
